services/api-gateway/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pymongo==4.6.1
motor==3.3.2
redis==5.0.1
aio-pika==9.3.1
minio==7.2.0
prometheus-client==0.19.0
python-i18n==0.3.9
python-dotenv==1.0.0
httpx==0.25.2
tenacity==8.2.3